LayoutTests/fast/events/touch/ios/touch-event-after-document-open-expected.txt

This tests triggering touch event on an element after document.open. The element should receive touch event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didFireTouchStartEvent is true
PASS didFireTouchEndEvent is true
